Add loggers options
