+++ lib/index.js
@@ -1,4 +1,4 @@
-const { logger } = require('./logger');
+const { logger } = require('./loggers');
 const { expressMiddleware, expressRequestIdMiddleware } = require('./middlewares');
 const { setRequestId, getRequestId, ns } = require('./namespace');
