+++ lib/middlewares.js
@@ -1,9 +1,8 @@
 const onFinished = require('on-finished');
 
 const { ns, setRequestId } = require('./namespace');
-const { logger } = require('./logger');
 
-const expressMiddleware = (req, res, next) => {
+const expressMiddleware = logger => (req, res, next) => {
   const { method, params, query, body } = req;
   const url = req.originalUrl || req.url;
