+++ test/logger.spec.js
@@ -13,15 +13,17 @@ jest.doMock('pino', () => () => ({
 }));
 const { logger } = require('..');
 
+const testLogger = logger();
+
 describe('logger', () => {
   test('should call pino when logger is called', () => {
-    logger.info('test');
+    testLogger.info('test');
     expect(loggerMock).toHaveBeenCalled();
     const msg = loggerMock.mock.calls[0][1];
     expect(msg).toBe('test');
   });
   test('should log id when requestId exists', () => {
-    logger.info('test');
+    testLogger.info('test');
     expect(loggerMock).toHaveBeenCalled();
     const id = loggerMock.mock.calls[0][0].trim();
     expect(id).toBe('[1]');
